fix types in format string
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@5510 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/roken/emalloc.c
@@ -56,6 +56,6 @@ emalloc (size_t sz)
     void *tmp = malloc (sz);
 
     if (tmp == NULL && sz != 0)
-	err (1, "malloc %u", sz);
+	err (1, "malloc %lu", (unsigned long)sz);
     return tmp;
 }
