+++ lib/roken/erealloc.c
@@ -56,6 +56,6 @@ erealloc (void *ptr, size_t sz)
     void *tmp = realloc (ptr, sz);
 
     if (tmp == NULL && sz != 0)
-	err (1, "realloc %u", sz);
+	err (1, "realloc %lu", (unsigned long)sz);
     return tmp;
 }
